Checkout code from private git repositories using ssh agent forwarding.
Currently when using the playbook with a private repository the easiest
way to do it is to copy the private key to the remote server. This is
extremely dangerous. If you lose control of the server then you have lost
control of your source code. In any reasonably busy project tracking down
changes from a malicious user will be very difficult indeed.

The solution is to enable SSH agent forwarding on the remote server (and
the Vagrant box) so when the code is checked out from a git repository,
to which an SSH public key has been added, the request to unlock the key
will be forwarded to the control node where the playbook (or Vagrant)
is being run.

Enabling SSH agent forwarding is simple. Vagrant supports it using the
configuration flag, config.ssh.forward_agent. Ansible also supports it
through the --extra-ssh-args which can be specified on the command line
when the playbook is run. You could also set it in ansible.cfg or as the
ansible_ssh_extra_args variable which could be set in the inventory or
playbook files. The command line option was chosen since it is the simplest
and does not interfere with any other options the user might want to set.

There were two complications when getting forwarding to work on the remote
server:

1. The git command is run as a separate process and so as a separate user.
   That meant the socket connection set up to forward the key unlock
   request was "lost". The solution was to add a task to preserve the
   SSH_AUTH_SOCK environment variable when the git module is run so it
   can still access forwarded ports. The task is conditional on the
   variables, setup_git_repo and ssh_forward_agent and the git tag was
   added so it only runs when git is being configured.

2. The ansible git module uses python's tempfile.mkstemp() which in turn
   uses tempfile.tempdir. The value is taken from the environment variable
   TMPDIR, and typically is set to /tmp. Unfortunately /tmp is mounted as
   noexec and so the command fails. The solution is to redefine the value
   of TMPDIR for the task to somethis with exec privileges, in this case
   /var/tmp, so the command can complete.

 The changes were made to roles/web/tasks/setup_git_repo.yml and are
 fully commented with linked to the issues so the rationle for the changes
 are clear to users.